Revert "CON-86-story(resolve multiple responses): enable admin resolve/unresolve multiple responses at once (#445)"
This reverts commit e366b37d8e491f2356bd08ebc7bf8d21c1d131c0.
