+++ fixtures/response/room_response_fixture.py
@@ -41,23 +41,6 @@ get_room_response_query_data = {
     }
 }
 
-get_resolved_room_responses_query = '''{
-    roomResponse(roomId:1, resolved:true) {
-        roomName,
-        totalResponses
-    }
-}
-'''
-
-get_resolved_room_responses_query_data = {
-    "data": {
-        "roomResponse": {
-            "roomName": "Entebbe",
-            "totalResponses": 1
-        }
-    }
-}
-
 get_room_response_query_by_date = '''
 query{{
     allRoomResponses(startDate: "{}",
@@ -302,60 +285,6 @@ summary_room_response_data = {
   }
 }
 
-all_resolved_room_response_query = '''{
-    allRoomResponses(resolved:true) {
-        responses {
-        roomName,
-        totalResponses,
-        response {
-          responseId,
-          response {
-            ... on Rate{
-              rate
-            }
-            ... on SelectedOptions{
-              options
-            }
-            ... on TextArea{
-              suggestion
-            }
-            ... on MissingItems{
-              missingItems{
-                name
-                id
-              }
-            }
-          }
-        }
-      }
-    }
-}
-'''
-
-all_resolved_room_response_data = {
-  "data": {
-    "allRoomResponses": {
-      "responses": [
-        {
-          "roomName": "Entebbe",
-          "totalResponses": 1,
-          "response": [
-            {
-              "responseId": 2,
-              "response": {
-                "options": [
-                  "marker pen",
-                  "apple tv"
-                ]
-              }
-            }
-          ]
-        }
-      ]
-    }
-  }
-}
-
 filter_by_response_query = '''
 query{
     allRoomResponses(upperLimitCount: 3, lowerLimitCount: 0 ){
@@ -489,61 +418,6 @@ filter_by_response_data = {
             ]
         }
     }
-}
-
-search_resolved_responses_by_room_name = '''
-query{
-    allRoomResponses(room:"Entebbe", resolved:true){
-        responses{
-            totalResponses
-            roomName
-            response{
-                responseId
-                response {
-                  ... on Rate{
-                    rate
-                  }
-                  ... on SelectedOptions{
-                    options
-                  }
-                  ... on TextArea{
-                    suggestion
-                  }
-                  ... on MissingItems{
-                    missingItems{
-                      name
-                      id
-                    }
-                  }
-                }
-            }
-        }
-    }
-}
-'''
-
-search_resolved_responses_by_room_name_data = {
-  "data": {
-    "allRoomResponses": {
-      "responses": [
-        {
-          "roomName": "Entebbe",
-          "totalResponses": 1,
-          "response": [
-            {
-              "responseId": 2,
-              "response": {
-                "options": [
-                  "marker pen",
-                  "apple tv"
-                ]
-              }
-            }
-          ]
-        }
-      ]
-    }
-  }
 }
 
 query_paginated_responses = '''
@@ -739,120 +613,3 @@ mark_a_user_response_as_unresolved_mutation_response = {
         }
     }
 }
-
-mark_multiple_as_resolved_mutation = '''
-    mutation{
-  resolveMultipleResponses(responses:[1], resolved:true){
-    handledResponses{
-      responses{
-        id
-        resolved
-        response{
-          ... on Rate{
-            rate
-          }
-          ... on TextArea{
-            suggestion
-          }
-          ... on SelectedOptions{
-            options
-          }
-          ... on MissingItems{
-            missingItems{
-              name
-            }
-        }
-        }
-      }
-      totalResponses
-    }
-  }
-}
-'''
-
-mark_multiple_as_resolved_response = {
-  "data": {
-    "resolveMultipleResponses": {
-      "handledResponses": {
-        "responses": [
-          {
-            "id": 1,
-            "resolved": True,
-            "response": {
-              "rate": 1
-            }
-          }
-        ],
-        "totalResponses": 1
-      }
-    }
-  }
-}
-
-mark_multiple_as_unresolved_mutation = '''
-    mutation{
-  resolveMultipleResponses(responses:[1], resolved:false){
-    handledResponses{
-      responses{
-        id
-        resolved
-        response{
-          ... on Rate{
-            rate
-          }
-          ... on TextArea{
-            suggestion
-          }
-          ... on SelectedOptions{
-            options
-          }
-          ... on MissingItems{
-            missingItems{
-              name
-            }
-        }
-        }
-      }
-      totalResponses
-    }
-  }
-}
-'''
-
-mark_multiple_as_unresolved_response = {
-  "data": {
-    "resolveMultipleResponses": {
-      "handledResponses": {
-        "responses": [
-          {
-            "id": 1,
-            "resolved": False,
-            "response": {
-              "rate": 1
-            }
-          }
-        ],
-        "totalResponses": 1
-      }
-    }
-  }
-}
-
-mark_multiple_as_resolved_with_invalid_id = '''
-    mutation{
-  resolveMultipleResponses(responses:[36], resolved:true){
-    handledResponses{
-      responses{
-        id
-        resolved
-        response{
-          ... on TextArea{
-          suggestion
-        }
-        }
-      }
-      totalResponses
-    }
-  }
-}
-'''
